pyproject.toml
[tool.poetry]
name = "doki"
version = "0.1.0"
description = "A tool to chat with technical documentation using RAG."
authors = ["MBoulahtouf <marouan.boulahtouf@protonmail.ch>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain = ">=0.2.1,<0.3.0"
langchain-community = ">=0.2.1,<0.3.0"
beautifulsoup4 = ">=4.12.3,<5.0.0"
sentence-transformers = ">=3.0.1,<4.0.0"
chromadb = ">=0.5.3,<0.6.0"
streamlit = ">=1.36.0,<2.0.0"
langchain-huggingface = ">=0.0.3,<0.1.0"
langchain-groq = ">=0.1.6"
python-dotenv = ">=1.0.1,<2.0.0"
langchain-chroma = "0.1.2"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
lxml = "^5.4.0"
selenium = "^4.33.0"
scrapy = "^2.13.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

